refactor: project-related services into separate packages
